pyproject.toml
[project]
name = "kbparse"
version = "0.1.0"
description = "PDF-first company knowledge-base parser with Markdown and RAG chunk outputs"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "typer>=0.12",
  "rich>=13.0",
  "python-dotenv>=1.0",
  "pillow>=10.0",
  "pymupdf>=1.24"
]

[project.optional-dependencies]
docling = ["docling>=2.0"]

[project.scripts]
kbparse = "kbparse.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kbparse"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
